refactor: renamed Streamregistry artifacts to V3
