+++ packages/client/src/StreamRegistry.ts
@@ -1,7 +1,7 @@
 import { Contract, ContractTransaction } from '@ethersproject/contracts'
 import { Signer } from '@ethersproject/abstract-signer'
-import type { StreamRegistry as StreamRegistryContract } from './ethereumArtifacts/StreamRegistry.d'
-import StreamRegistryArtifact from './ethereumArtifacts/StreamRegistryAbi.json'
+import type { StreamRegistryV3 as StreamRegistryContract } from './ethereumArtifacts/StreamRegistryV3'
+import StreamRegistryArtifact from './ethereumArtifacts/StreamRegistryV3Abi.json'
 import { BigNumber } from '@ethersproject/bignumber'
 import { Provider } from '@ethersproject/providers'
 import { scoped, Lifecycle, inject, DependencyContainer } from 'tsyringe'
