+++ packages/client/src/ethereumArtifacts/StreamRegistryV3.d.ts
