+++ packages/client/src/ethereumArtifacts/StreamRegistryV3Abi.json
